R3 - moving Header and Modals to pages folder
