+++ src/components/CustomModal/index.ts
@@ -1,0 +1,3 @@
+import CustomModal from "./CustomModal";
+
+export default CustomModal;
