+++ src/components/CustomModal/styled.tsx
